* base/pidfile.c (pidfile_remove): Make pidfile removal more robust. The contents of the pidfile are now checked before removal to prevent accidental deletion of the file by other processes. Removed now obsolete debug message.
git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/trunk/openvas-libraries@5959 423fd1db-d629-0410-8442-d21db03e70f4

+++ base/pidfile.c
@@ -87,10 +87,20 @@ pidfile_remove (gchar * daemon_name)
 {
   gchar *name_pid = g_strconcat (daemon_name, ".pid", NULL);
   gchar *pidfile_name = g_build_filename (OPENVAS_PID_DIR, name_pid, NULL);
+  gchar *pidfile_contents;
 
   g_free(name_pid);
 
-  g_unlink (pidfile_name);
+  if (g_file_get_contents (pidfile_name, &pidfile_contents, NULL, NULL))
+    {
+      int pid = atoi (pidfile_contents);
+
+      if (pid == getpid ())
+        {
+          g_unlink (pidfile_name);
+        }
+      g_free (pidfile_contents);
+    }
+
   g_free (pidfile_name);
-  g_debug ("   Deleted pidfile for %s.\n", daemon_name);
 }
